moved dialogue and rendering widgets to a new file
